Close sesh with defer
This commit adds a way to close the session of dmsg via a defer statement with a go routine in it inside the fun RoundTrip().

The goroutine checks if the response body is closed every 10 seconds and if it is closed then we clsoe the session if not then we keep checking till it's closed or if context is closed.